Revert "pmt-ptw"
This reverts commit 28c7b49aa17a325bd46f34047003c48f0137ee99.

+++ src/main/java/io/reactivex/rxjava3/internal/jdk8/ParallelCollector.java
@@ -98,7 +98,7 @@ public final class ParallelCollector<T, A, R> extends Flowable<R> {
                     }
                 }
 
-                int c = ((Integer)(curr.tryAcquireSlot())).intValue();
+                int c = curr.tryAcquireSlot();
                 if (c < 0) {
                     current.compareAndSet(curr, null);
                     continue;
