+++ src/main/java/io/reactivex/rxjava3/internal/operators/parallel/ParallelFlatMap.java
@@ -15,7 +15,6 @@ package io.reactivex.rxjava3.internal.operators.parallel;
 
 import org.reactivestreams.*;
 
-import io.reactivex.rxjava3.core.myBlackhole;
 import io.reactivex.rxjava3.functions.Function;
 import io.reactivex.rxjava3.internal.operators.flowable.FlowableFlatMap;
 import io.reactivex.rxjava3.parallel.ParallelFlowable;
@@ -68,7 +67,7 @@ public final class ParallelFlatMap<T, R> extends ParallelFlowable<R> {
         final Subscriber<T>[] parents = new Subscriber[n];
 
         for (int i = 0; i < n; i++) {
-            parents[i] = FlowableFlatMap.subscribe(subscribers[((Integer)(i)).intValue()], mapper, delayError, maxConcurrency, prefetch);
+            parents[i] = FlowableFlatMap.subscribe(subscribers[i], mapper, delayError, maxConcurrency, prefetch);
         }
 
         source.subscribe(parents);
