+++ src/main/java/io/reactivex/rxjava3/internal/subscribers/BasicFuseableConditionalSubscriber.java
@@ -15,7 +15,6 @@ package io.reactivex.rxjava3.internal.subscribers;
 
 import org.reactivestreams.Subscription;
 
-import io.reactivex.rxjava3.core.myBlackhole;
 import io.reactivex.rxjava3.exceptions.Exceptions;
 import io.reactivex.rxjava3.internal.fuseable.*;
 import io.reactivex.rxjava3.internal.subscriptions.SubscriptionHelper;
@@ -134,7 +133,7 @@ public abstract class BasicFuseableConditionalSubscriber<T, R> implements Condit
         QueueSubscription<T> qs = this.qs;
         if (qs != null) {
             if ((mode & BOUNDARY) == 0) {
-                int m = qs.requestFusion(((Integer)(mode)).intValue());
+                int m = qs.requestFusion(mode);
                 if (m != NONE) {
                     sourceMode = m;
                 }
